Refactor: Enhance Modularity with New Downloader and Assembler Packages
I've restructured the code to bolster modularity and readability. The functions DownloadPartFiles and AssembleFileFromParts have been relocated into their own downloader and assembler packages, respectively.

This move enhances the separation of concerns and encapsulation in my codebase, facilitating future maintenance and updates.

Moreover, I've adjusted the function calls in the main application to utilize the new assembler.NewAssembler() and downloader.NewDownloader() methods, ensuring smooth integration with these new packages.
